+++ langchain/chains/prompt_selector.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, Field
 from langchain.chat_models.base import BaseChatModel
 from langchain.llms.base import BaseLLM
 from langchain.prompts.base import BasePromptTemplate
-from langchain.schema import BaseLanguageModel
+from langchain.base_language_model import BaseLanguageModel
 
 
 class BasePromptSelector(BaseModel, ABC):
