+++ langchain/memory/kg.py
@@ -12,7 +12,8 @@ from langchain.memory.prompt import (
 )
 from langchain.memory.utils import get_buffer_string, get_prompt_input_key
 from langchain.prompts.base import BasePromptTemplate
-from langchain.schema import BaseLanguageModel, SystemMessage
+from langchain.schema import SystemMessage
+from langchain.base_language_model import BaseLanguageModel
 
 
 class ConversationKGMemory(BaseChatMemory, BaseModel):
