+++ langchain/memory/summary.py
@@ -7,7 +7,8 @@ from langchain.memory.chat_memory import BaseChatMemory
 from langchain.memory.prompt import SUMMARY_PROMPT
 from langchain.memory.utils import get_buffer_string
 from langchain.prompts.base import BasePromptTemplate
-from langchain.schema import BaseLanguageModel, BaseMessage, SystemMessage
+from langchain.schema import BaseMessage, SystemMessage
+from langchain.base_language_model import BaseLanguageModel
 
 
 class SummarizerMixin(BaseModel):
